extracted method to new service.
